Fixing Heroku issues

+++ db/migrate/20121031150716_add_price_to_payment.rb
@@ -1,5 +1,5 @@
 class AddPriceToPayment < ActiveRecord::Migration
   def change
-    add_column :payments, :price, :double
+    add_column :payments, :price, :float
   end
 end
